Week 02: Lab 2 (Fix DebugExercise1.divideThenRound

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,8 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double) top / (double) bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
@@ -22,7 +22,7 @@ public class DebugExercise1 {
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
-        int t3 = 3;
+            int t3 = 3;
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
